backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
alembic==1.13.3
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
psycopg2-binary==2.9.9
